requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1  # For parquet files
python-dotenv>=1.0.0

# Machine Learning & Embeddings
sentence-transformers==2.2.2
scikit-learn==1.3.2
chromadb>=0.4.0

# Vector Database
chromadb==0.4.18

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Progress Bars
tqdm==4.66.1

# Google Gemini API 
google-generativeai>=0.3.0

# Data generation
faker>=20.0.0

# Streamlit Dashboard 
streamlit==1.28.2

# Environment Variables
python-dotenv==1.0.0

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
nbformat==5.9.2

# For Anthropic
# anthropic==0.7.8

# Utilities
requests==2.31.0
urllib3==2.1.0